+++ mustang_core/lib/mustang_core.dart
@@ -20,6 +20,8 @@ export 'src/cache/mustang_cache.dart';
 export 'src/cache/wrench_cache.dart';
 export 'src/config/mustang_app_config.dart';
 export 'src/config/mustang_app_config.model.dart';
+export 'src/config/mustang_app_states.dart';
+export 'src/config/mustang_app_states.model.dart';
 export 'src/event_stream/event_stream.dart';
 export 'src/store/mustang_store.dart';
 export 'src/contracts/mustang_state.dart';