+++ mustang_core/lib/src/config/mustang_app_states.dart
@@ -1,0 +1,9 @@
+import 'package:built_collection/built_collection.dart';
+import 'package:mustang_core/mustang_core.dart';
+
+@appModel 
+abstract class $MustangAppStates {
+  @InitField([])
+  late BuiltList<String> states;
+}
+